refactor: standardize Fab component structure and naming
Restructured Fab component by separating logic, styles, and interfaces for better maintainability. Renamed files and updated exports to reflect a more consistent naming convention (`Fab` to `FabComponent`). Adjusted stories and updated related styles to align with the new component design.
